% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rearth.R
\name{Rearth}
\alias{Rearth}
\title{An animated map for vector data
This project is just at its very beginning!}
\usage{
Rearth(viewpoint = "", mode = "air", fn, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{viewpoint}{Text. Following the pattern "$projection=$x,$y,$z".}

\item{mode}{Currently only two values: "air" or "ocean".}

\item{fn}{filename.}

\item{width}{Width of the display area of the widget.}

\item{height}{Height of the display area of the widget.}

\item{elementId}{Name of the widget.}
}
\description{
An animated map for vector data
This project is just at its very beginning!
}
